+++ packages/kit/tests/fixtures/optimize-KjI6IT/.vite/deps_temp_685830c0/@implementjs_kit_params.js
@@ -1,0 +1,18 @@
+import {
+  appMatchers,
+  isParamMatcher,
+  matcher,
+  matcherTable,
+  mismatch,
+  registerMatchers
+} from "./chunk-R47OUD5L.js";
+import "./chunk-PX6F3LHL.js";
+export {
+  appMatchers,
+  isParamMatcher,
+  matcher,
+  matcherTable,
+  mismatch,
+  registerMatchers
+};
+//# sourceMappingURL=@implementjs_kit_params.js.map