+++ packages/lucide/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @implementjs/lucide
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`0ac0208`](https://github.com/ieedan/implement/commit/0ac0208d825804969a58c61fb21063724a10e431)]:
+  - @implementjs/core@0.0.8
+
 ## 0.0.7
 
 ### Patch Changes